math checkin refactor in progress
