+++ src/math/index.ts
@@ -23,7 +23,12 @@ SOFTWARE.*/
 
 export * from "./constants.js";
 export { type Mat4x4, Mat4x4Schema } from "./mat4x4/index.js";
-export { Vec2Schema, type Vec2 } from "./vec2/index.js";
+
+// Export the type Vec2
+export type Vec2 = import("./vec2/vec2.js").Type;
+// Export the namespace Vec2
+export * as Vec2 from "./vec2/vec2.js";
+
 export { Vec3Schema, type Vec3 } from "./vec3/index.js";
 export { Vec4Schema, type Vec4 } from "./vec4/index.js";
 export * from "./picking/index.js";
